+++ tests/mutations.rs
@@ -59,7 +59,7 @@ fn small_component() {
         .world_mut()
         .query::<&BoolComponent>()
         .single(client_app.world());
-    assert!(component.0, "changed value should be updated on client");
+    assert!(component.0, "mutated value should be updated on client");
 }
 
 #[test]
@@ -341,7 +341,7 @@ fn marker_with_history() {
     server_app.update();
     server_app.exchange_with_client(&mut client_app);
 
-    // Change value again to generate another update.
+    // Change value again to trigger another message.
     let mut component = server_app
         .world_mut()
         .get_mut::<BoolComponent>(server_entity)
@@ -357,8 +357,8 @@ fn marker_with_history() {
     assert_eq!(
         history.0,
         [false, false, true],
-        "the initial value should come first, then the latest update, \
-        and after that the older update because recent updates processed first"
+        "the initial value should come first, then the latest mutation, \
+        and after that the older mutation because recent mutations processed first"
     );
 }
 
@@ -418,38 +418,38 @@ fn marker_with_history_consume() {
     server_app.exchange_with_client(&mut client_app);
 
     // Change value, but don't process it on client.
-    let update_entity1 = server_app.world_mut().spawn_empty().id();
+    let dummy_entity1 = server_app.world_mut().spawn_empty().id();
     let mut component = server_app
         .world_mut()
         .get_mut::<MappedComponent>(server_entity)
         .unwrap();
-    component.0 = update_entity1;
+    component.0 = dummy_entity1;
 
     server_app.update();
     server_app.exchange_with_client(&mut client_app);
 
-    // Change value again to generate another update.
-    let update_entity2 = server_app.world_mut().spawn_empty().id();
+    // Change value again to trigger another message.
+    let dummy_entity2 = server_app.world_mut().spawn_empty().id();
     let mut component = server_app
         .world_mut()
         .get_mut::<MappedComponent>(server_entity)
         .unwrap();
-    component.0 = update_entity2;
+    component.0 = dummy_entity2;
 
     server_app.update();
     server_app.exchange_with_client(&mut client_app);
     client_app.update();
 
     let entity_map = client_app.world().resource::<ServerEntityMap>();
-    assert!(entity_map.to_client().contains_key(&update_entity2));
+    assert!(entity_map.to_client().contains_key(&dummy_entity2));
     assert!(
-        !entity_map.to_client().contains_key(&update_entity1),
-        "client should consume older update for other components with marker that requested history"
+        !entity_map.to_client().contains_key(&dummy_entity1),
+        "client should consume older mutations for other components with marker that requested history"
     );
     assert_eq!(
         client_app.world().entities().len(),
         3,
-        "client should have 2 initial entities and 1 from update"
+        "client should have 2 initial entities and 1 from mutate message"
     );
 }
 
@@ -503,7 +503,7 @@ fn marker_with_history_old_update() {
     server_app.exchange_with_client(&mut client_app);
 
     // Artificially make the last confirmed tick too large
-    // so that the next update for this entity is discarded.
+    // so that the next mutation for this entity is discarded.
     let mut tick = **server_app.world().resource::<ServerTick>();
     tick += u64::BITS + 1;
     let mut history = client_app
@@ -530,7 +530,7 @@ fn marker_with_history_old_update() {
     assert_eq!(
         history.0,
         [false],
-        "update should be considered too old and discarded"
+        "mutation should be considered too old and discarded"
     );
 }
 
@@ -704,7 +704,7 @@ fn with_despawn() {
         .unwrap();
     component.0 = true;
 
-    // Update without client to send update message.
+    // Update without client to send mutate message.
     server_app.update();
 
     server_app.world_mut().despawn(server_entity);
@@ -745,7 +745,7 @@ fn buffering() {
     client_app.update();
     server_app.exchange_with_client(&mut client_app);
 
-    // Artificially reset the init tick to force the next received update to be buffered.
+    // Artificially reset the init tick to force the next received mutation to be buffered.
     let mut init_tick = client_app.world_mut().resource_mut::<ServerInitTick>();
     let previous_tick = *init_tick;
     *init_tick = Default::default();
@@ -764,9 +764,9 @@ fn buffering() {
         .world_mut()
         .query::<&BoolComponent>()
         .single(client_app.world());
-    assert!(!component.0, "client should buffer the update");
+    assert!(!component.0, "client should buffer the mutation");
 
-    // Restore the init tick to let the buffered update apply
+    // Restore the init tick to let the buffered mutation apply
     *client_app.world_mut().resource_mut::<ServerInitTick>() = previous_tick;
 
     server_app.update();
@@ -777,7 +777,7 @@ fn buffering() {
         .world_mut()
         .query::<&BoolComponent>()
         .single(client_app.world());
-    assert!(component.0, "buffered update should be applied");
+    assert!(component.0, "buffered mutation should be applied");
 }
 
 #[test]
@@ -809,23 +809,23 @@ fn old_ignored() {
     server_app.exchange_with_client(&mut client_app);
 
     // Change the value, but don't process it on client.
-    let update_entity1 = server_app.world_mut().spawn_empty().id();
+    let dummy_entity1 = server_app.world_mut().spawn_empty().id();
     let mut component = server_app
         .world_mut()
         .get_mut::<MappedComponent>(server_entity)
         .unwrap();
-    component.0 = update_entity1;
+    component.0 = dummy_entity1;
 
     server_app.update();
     server_app.exchange_with_client(&mut client_app);
 
-    // Change the value again to generate another update.
-    let update_entity2 = server_app.world_mut().spawn_empty().id();
+    // Change the value again to trigger another message.
+    let dummy_entity2 = server_app.world_mut().spawn_empty().id();
     let mut component = server_app
         .world_mut()
         .get_mut::<MappedComponent>(server_entity)
         .unwrap();
-    component.0 = update_entity2;
+    component.0 = dummy_entity2;
 
     server_app.update();
     server_app.exchange_with_client(&mut client_app);
@@ -833,13 +833,13 @@ fn old_ignored() {
 
     let entity_map = client_app.world().resource::<ServerEntityMap>();
     assert!(
-        !entity_map.to_client().contains_key(&update_entity1),
-        "client should ignore older update"
+        !entity_map.to_client().contains_key(&dummy_entity1),
+        "client should ignore older mutation"
     );
     assert_eq!(
         client_app.world().entities().len(),
         3,
-        "client should have 2 initial entities and 1 from update"
+        "client should have 2 initial entities and 1 from mutation"
     );
 }
 
@@ -852,7 +852,7 @@ fn acknowledgment() {
             MinimalPlugins,
             RepliconPlugins.set(ServerPlugin {
                 tick_policy: TickPolicy::EveryFrame,
-                update_timeout: Duration::ZERO, // Will cause dropping updates after each frame.
+                mutate_timeout: Duration::ZERO, // Will cause dropping updates after each frame.
                 ..Default::default()
             }),
         ))
@@ -904,7 +904,7 @@ fn acknowledgment() {
 
     assert!(
         tick1.get() < tick2.get(),
-        "client should receive the same update twice because server missed the ack"
+        "client should receive the same mutation twice because server missed the ack"
     );
 
     server_app.update();
@@ -920,7 +920,7 @@ fn acknowledgment() {
     assert_eq!(
         tick2.get(),
         tick3.get(),
-        "client shouldn't receive acked update"
+        "client shouldn't receive acked mutation"
     );
 }